fixed hook, etc

+++ classes/Vector.cs
@@ -2,11 +2,11 @@
 /// @CSharpLua.Ignore
 /// </summary>
 public class Vector {
-    public int x;
-    public int y;
-    public int z;
+    public float x;
+    public float y;
+    public float z;
 
-    public Vector(int _x, int _y, int _z) {
+    public Vector(float _x, float _y, float _z) {
         x = _x;
         y = _y;
         z = _z;
